Type check as a composite project
This does *not* get tests passing in the new format, but *does* get
the whole project 'working' end to end otherwise.
