+++ packages/result/package.json
@@ -1,6 +1,15 @@
 {
   "name": "@true-myth/result",
   "version": "5.0.0",
+  "type": "module",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "exports": {
+    ".": "./dist/index.js",
+    "./*": "./dist/*",
+    "./package.json": "./package.json"
+  },
+  "types": "./dist/index.d.ts",
   "dependencies": {
     "@true-myth/utils": "5.0.0"
   },
