+++ packages/true-myth/package.json
@@ -21,13 +21,12 @@
   },
   "version": "5.0.0",
   "type": "module",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
   "exports": {
     ".": "./dist/index.js",
-    "./maybe": "./dist/maybe.js",
-    "./result": "./dist/result.js",
-    "./unit": "./dist/unit.js"
+    "./*": "./dist/*",
+    "./package.json": "./package.json"
   },
   "types": "index.d.ts",
   "keywords": [
@@ -48,17 +47,6 @@
     "@true-myth/maybe": "5.0.0",
     "@true-myth/result": "5.0.0",
     "@true-myth/utils": "5.0.0"
-  },
-  "devDependencies": {
-    "prettier": "^2.2.1",
-    "prettier-eslint": "^13.0.0",
-    "release-it": "^14.11.8",
-    "release-it-lerna-changelog": "^4.0.1",
-    "rimraf": "^3.0.2",
-    "shelljs": "^0.8.2",
-    "ts-jest": "^27.0.7",
-    "typedoc": "^0.22.10",
-    "typescript": "~4.5.2"
   },
   "volta": {
     "extends": "../../package.json"
