+++ packages/true-myth/src/maybe.ts
@@ -1,1 +1,1 @@
-export {};
+export * from '@true-myth/maybe';
