+++ packages/true-myth/src/unit.ts
@@ -1,0 +1,1 @@
+export { Unit } from '@true-myth/result';
