+++ ts/es.tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../tsconfig.json",
   "include": ["../src"],
   "compilerOptions": {
-    "module": "ESNext",
     "declaration": false,
     "outDir": "../dist"
   }
